requirementspypy.txt
Django
amqp
anyjson
Babel
billiard
celery
certifi
chardet
coreapi
coreschema
django-celery
django-cors-headers
django-crispy-forms
django-filter
django-formtools
django-redis
django-reversion
djangorestframework
djangorestframework-jwt
drf-extensions
et-xmlfile
flower
future
gunicorn
httplib2
idna
itypes
jdcal
Jinja2
kombu
Markdown
MarkupSafe
mysqlclient
openpyxl
Pillow
pycryptodome
PyJWT
pytz
raven
redis
requests
six
tornado
uritemplate
urllib3
uWSGI
vine
xlrd
XlsxWriter
xlwt
yunpian-python-sdk
